move SeedsESR utility class to v2, remove duplicares
