+++ lib/v2/blocs/deeplink/usecase/get_initial_deep_link_use_case.dart
@@ -1,9 +1,9 @@
 import 'package:async/async.dart';
 
 // ignore: import_of_legacy_library_into_null_safe
-import 'package:seeds/features/scanner/telos_signing_manager.dart';
 import 'package:seeds/v2/blocs/deeplink/model/deep_link_data.dart';
 import 'package:seeds/v2/datasource/local/settings_storage.dart';
+import 'package:seeds/v2/datasource/local/util/seeds_esr.dart';
 
 class GetInitialDeepLinkUseCase {
   Future<Result> run(Uri newLink) async {
