+++ lib/v2/datasource/local/qr_code_service.dart
@@ -2,6 +2,7 @@ import 'package:async/async.dart';
 // ignore: import_of_legacy_library_into_null_safe
 import 'package:dart_esr/dart_esr.dart';
 import 'package:seeds/v2/datasource/local/models/scan_qr_code_result_data.dart';
+import 'package:seeds/v2/datasource/local/util/seeds_esr.dart';
 
 class QrCodeService {
   Future<Result<dynamic>> processQrCode(String scanResult, String accountName) {
@@ -12,7 +13,7 @@ class QrCodeService {
       return Future.value(ErrorResult('Invalid QR Code'));
     }
 
-    final _SeedsESR esr = _SeedsESR(uri: scanResult);
+    final SeedsESR esr = SeedsESR(uri: scanResult);
     return esr.resolve(account: accountName).then((value) => processResolvedRequest(esr)).catchError((onError) {
       print(" processQrCode : Error processing QR code");
       return ErrorResult("Error processing QR code");
@@ -20,7 +21,7 @@ class QrCodeService {
   }
 }
 
-Result processResolvedRequest(_SeedsESR esr) {
+Result processResolvedRequest(SeedsESR esr) {
   final Action action = esr.actions.first;
   if (_canProcess(action)) {
     final Map<String, dynamic> data = Map<String, dynamic>.from(action.data! as Map<dynamic, dynamic>);
@@ -35,37 +36,4 @@ Result processResolvedRequest(_SeedsESR esr) {
 
 bool _canProcess(Action action) {
   return action.account!.isNotEmpty && action.name!.isNotEmpty;
-}
-
-class _SeedsESR {
-  late SigningRequestManager manager;
-
-  late List<Action> actions;
-
-  _SeedsESR({String? uri}) {
-    manager = _TelosSigningManager.from(uri);
-  }
-
-  Future<void> resolve({String? account}) async {
-    actions = await manager.fetchActions(account: account);
-  }
-}
-
-extension _TelosSigningManager on SigningRequestManager {
-  static SigningRequestManager from(String? uri) {
-    return SigningRequestManager.from(uri,
-        options: defaultSigningRequestEncodingOptions(nodeUrl: 'https://api.eos.miami'));
-  }
-
-  Future<List<Action>> fetchActions({String? account, String permission = "active"}) async {
-    final abis = await fetchAbis();
-
-    final auth = Authorization();
-    auth.actor = account;
-    auth.permission = permission;
-
-    final actions = resolveActions(abis, auth);
-
-    return actions;
-  }
 }
