+++ lib/v2/screens/app/interactor/mappers/singing_request_state_mapper.dart
@@ -1,5 +1,5 @@
 // ignore: import_of_legacy_library_into_null_safe
-import 'package:seeds/features/scanner/telos_signing_manager.dart';
+import 'package:seeds/v2/datasource/local/util/seeds_esr.dart';
 import 'package:seeds/v2/domain-shared/page_command.dart';
 import 'package:seeds/v2/domain-shared/page_state.dart';
 import 'package:seeds/v2/domain-shared/result_to_state_mapper.dart';
